Export useful client test classes
Moves `FakeClient`, `FakeService`, the dispatch generator and the
utility function `to_create_params` to the test module.
