+++ src/frequenz/client/dispatch/test/__init__.py
@@ -1,0 +1,4 @@
+# License: MIT
+# Copyright © 2024 Frequenz Energy-as-a-Service GmbH
+
+"""Helpful utilities for testing with the dispatch client."""
